wait for packages, bump all to 1.25.0
https://coveord.atlassian.net/browse/CDX-828
