+++ package.json
@@ -32,7 +32,7 @@
     "eslint-config-prettier": "8.3.0",
     "eslint-plugin-node": "11.1.0",
     "eslint-plugin-prettier": "4.0.0",
-    "exponential-backoff": "3.1.0",
+    "exponential-backoff": "^3.1.0",
     "fs-extra": "10.0.0",
     "gts": "3.1.0",
     "husky": "7.0.4",
