+++ packages/cli/package.json
@@ -39,10 +39,10 @@
   "devDependencies": {
     "@amplitude/types": "1.9.2",
     "@babel/core": "7.17.2",
-    "@coveo/angular": "1.24.1",
+    "@coveo/angular": "1.25.0",
     "@coveo/cra-template": "1.25.0",
     "@coveo/create-atomic": "1.25.0",
-    "@coveo/vue-cli-plugin-typescript": "1.24.1",
+    "@coveo/vue-cli-plugin-typescript": "1.25.0",
     "@oclif/dev-cli": "1.26.10",
     "@oclif/errors": "1.3.5",
     "@oclif/test": "1.2.9",
