+++ packages/vue-cli-plugin-typescript/package.json
@@ -16,7 +16,7 @@
   },
   "devDependencies": {
     "@coveo/headless": "1.42.0",
-    "@coveo/search-token-server": "1.24.1",
+    "@coveo/search-token-server": "1.25.0",
     "buefy": "0.9.14",
     "dotenv": "16.0.0",
     "vue": "3.2.30",
