+++ scripts/releaseV2/utils/wait-for-published-packages.js
@@ -1,0 +1,23 @@
+const {backOff} = require('exponential-backoff');
+const {spawnSync} = require('child_process');
+
+function showPackage(packageName) {
+  spawnSync('npm', ['show', packageName]);
+}
+
+export async function waitForPackage(packageName) {
+  try {
+    const pollInterval = 1e3;
+    await backOff(() => showPackage(packageName), {
+      delayFirstAttempt: true,
+      startingDelay: pollInterval,
+      maxDelay: pollInterval * 30,
+    });
+  } catch (err) {
+    console.error(`Package ${packageName} does not exist`);
+  }
+}
+
+export function waitForPackages(packagesToWait) {
+  return Promise.all(packagesToWait.map(waitForPackage));
+}
